Find matching distribution for combined hashed spec
Some hash join operations, that contained outer hash join as their inner part,
could wrongly enforce data redistribution only for the outer part, what lead
to inconsistent join result.

During [CPhysical*Join] optimization PdshashedMatching function could wrongly
interpret the combined hashed distribution spec derived for the outer join,
which was the inner child of the [CPhysical*Join] expression being optimized.
The combined hashed spec for the inner child (like hash left join in the test)
is derived from both outer and inner relations (it's done in PdsDerive function
of CPhysicalLeftOuterHashJoin), i.e. the derived distribution spec for the outer
join consists of 2 CDistributionSpecHashed entries, one for
each relation. Because join is outer and NULLs are only added to unmatched rows,
distribution spec for the inner relation has NullsColocated flag set to false.
And those 2 entries are represented as linked list, where the inner spec goes
first. Because of that, when we try to find matching distribution for such
combined hashed spec inside the PdshashedMatching function, we could build a
distribution matching the first spec from combined spec (one that has
Nullscolocated false) judging only by hash exprs. This logic is valid only
for the cases when the matching distribution does not require NULLs to be
co-located (NullsColocated false). And in cases when matching distribution spec
requires co-located NULLs this logic could lead to a discrepancy between
parts of join expression being optimized, because inner distribution has
NullsColocated false, but the outer (matching) distribution is required to have
NullsColocated to be true (taking into account that the hash exprs are
equivalent).

This patch prevents distribution mismatch for described case in the following
way: In case when matching distribution has NullsColocated set to true but the
input distribution has NullsColocated set to false, we try to build a matching
distribution spec based on input's equivalent distribution (pdshashedEquiv). If
none of the equivalent specs from combined distribution satisfies the condition
of NULLs colocation, we eventually raise the exception.