Fix SIGSEGV while executing lateral sub-queries with aggregate functions
Problem description:
Some queries with a lateral join cause a segmentation fault. It happens if the
query has a sub-query with aggregate functions and without a groupby clause.

Cause:
Postgres planner for these queries creates a plan with slices, and the lateral
variable is accessed from a different slice. Lateral variables can't be sent
across motions. So, it causes a segmentation fault.

Fix:
Fix is to force postgres planner to bring such queries to a single query
executor by setting 'force_singleQE' in the PlannerConfig structure. Previously,
this flag was set only for queries with lateral variables and a sub-query with
groupby / limit / distinct clauses. Now this condition is complemented by
checking the presence of aggregate functions.

It is rework of previous commit ff57098d0963fa40cf1af2ae7d4fb6d8d229250f