Make frozen_insert_crash test stable
Commit 8accc02 introduced a new flaking test. The test expects a panic to occur
in the process that performs the insert. But sometimes the panic happens first
in process which started by dtx recovery process that periodically checks for
orphaned transactions. Which leads to an error in the test. The solution is to
panic only the required process, which is done by extracting session from
pg_stat_activity by query.

Cherry-pick of 3d4cc61b6d697e7eb524078285d8dfbb19949e57