Fix dereference a potentially null pointer in atpxPart_validate_spec
Problem:
Static code analyzer found a potential issue with 'pNode'/'pNode2' pointer, that
is dereferenced before checking for a NULL value.

Fix:
1. Assert is added at the beginning of the function to ensure that 'pNode' and
the accessed fields of 'pNode' are not NULL.
2. Assert is added at the beginning of the loop to ensure that the accessed
fields of 'pNode2' are not NULL.
3. Redundant check for 'pNode2' not being NULL is removed, as it is checked by
the 'while' condition.
4. Assert is added after the update of 'pNode2' to ensure that the accessed
fields of the 'pNode2' are not NULL.
5. Second redundant check for 'pNode2' not being NULL is removed, as it is
guaranteed above by statements 'if (prule && prule->children)' and
'pNode2 = prule->children;'.