Fix an incorrect exit code in case of failure of resource group tests.
If the tests of the resource groups fail, the script should return a non-zero
exit code. But bash returns an exit code as a result of the last command; this
led to that script always exiting with a zero exit code.

This patch adds strict mode to the bash script and makes the shutdown of
containers a separate function that will be called on exit. The command that run
the test is now the last command, which ensures the script will return its exit
code.