Fix potential NULL pointer dereferencing in doDispatchDtxProtocolCommand
Problem:
In the 'doDispatchDtxProtocolCommand' function, the variable 'resultCount' was
declared, but not initialized. As it is an automatic variable, its value was
undefined. The pointer to 'resultCount' is passed to
'CdbDispatchDtxProtocolCommand' function as 'numresults'.
'CdbDispatchDtxProtocolCommand' didn't set the '*numresults' to 0 in case it
returned NULL. This return value is assigned to 'results'. So, if the 'results'
was NULL, 'resultCount' could have a non-zero value. And in this case, 'results'
could be dereferenced.

Fix:
Function 'doDispatchDtxProtocolCommand' is changed to immediately return false
if 'results' is NULL. It is equivalent to previous intended behavior, as for
empty 'results', the 'resultCount' is supposed to be 0. Plus 'resultCount' is
now initialized to 0, and CdbDispatchDtxProtocolCommand' additionally sets
'*numresults' to 0 in case it returns NULL.