Fix 'adb_collect_table_stats' failure
Problem:
Function 'adb_collect_table_stats' failed to execute after creation of a table
that is partitioned by a column with a non-timestamp type (particularly, by the
'bigint' column).

Cause:
There is a query in 'adb_collect_table_stats', which checks if there is
presented a partition of 'db_files_history' that covers today's date. When
this query was planned, the filter that checks the partitioning was planned to
execute before filtering by the relname. Thus the query tried to cast 'bigint'
to 'timestamp', which led to an error.

Fix:
The query is rewritten in a way that the filter that checks the partitioning is
executed after all other filters.