Implement default transformation flag for gpfdist
Currently, gpfdist transformation can only be applied to table source on
external table level. It can be done by specifying #transform tag. Sometimes it
may be not handy to specify the same transformation for all source files. The
goal of the patch is to implement a new flag with the default transformation
name. The default transformation can be set at gpfdist start and applied only if
table level rule is not specified.

New flag is a subject of only basic check. The idea was to keep the same
behavior and check rule existence only at runtime.
The test shows a behavior described in patch goal. Writable tables are not
considered here as their behavior is the same.