Fix segfault when copying dropped column ACLs to a new partition
While copying columns ACLs when adding a new partition, we assumed that the
attribute numbers in the source and target relation would match, which is not
true in the case of the dropped columns in the original relation. In this case,
we would try to get an ACL for a non-existent column in the target table, which
would lead to a segfault.

This patch adds a separate iterator for the list of columns in the target
relation to obtain the correct skipping of dropped columns in the original
relation and obtain the correct attribute number for the target relation. We
assume that there are no dropped columns in the target relation, since this is a
newly created relation.