Fix error at select from replicated table in function
For replicated tables there is adjusting, which divides count of processed
tuples on number of segments at QD at the end of query execution for DML
operations. Such adjusting was applied, when we do select from function, which
contains DML operation with replicated table. It was incorrect behavior, because
the main operation is SELECT.

This patch add check, which allows not to do adjusting if current query is not
DML.