Updated approach to counting the number of rows for replicated tables at DML
Replicated tables are the same at all segments. Modifying replicated tables lead
to do the same processing at each segment. All segments calculate the count of
modified (updated, deleted or inserted) rows. The count on all segments will be
the same. It led to situation: QD got the same count of modified rows from each
segments, summarize it, then divide it on number of segments. But it had to do
not in all cases. One such case: select from a function which contains DML on
replicated table. Such query led to an error: "consistency check on SPI tuple
count failed".

Also, the ExecutePlan function allows to limit the number of processing tuples,
but for DML operations the number of tuples is get from summ of such numbers
from segments. It led to the same error, because QD processed the number of
tuples, which is less than all segments sent.

This patch modifies approach to counting the number of rows for replicated
tables at DML operations. Only one segment will calculate the count of modified
rows if the table is replicated. Other segments will send zero. Dividing on
number of segments is no longer needed. Also, saving the number of current
processed tuples to the all processed tuples was added if processing tuples was
broken by the limit.