Materialize non-SELECT CTEs constantly
Planner inlined each reference to CTE containing modifying DML operations into
the query plan. This inlining led to several problems. Each reference to the CTE
resulted in the operation being repeated, causing duplicating data modifications
That behaviour could lead to results that are inconsistent or erroneous.

The proposed solution involves constant materialization of such CTEs using the
SharedScan node, which functions similarly to the CTEScan node in vanilla
PostgreSQL (in local SharedInputScan case). This approach prevents the
repeated execution of the modifying operations by sharing the results of the CTE
across different parts of the query plan. To implement this change, a
modification in the set_cte_pathlist function is done. Specifically, if the
subquery's command type is not SELECT, the CTE is marked for sharing.