Add collection of test logs of resource groups.
During the testing of resource groups, log files are created that can be useful
in analyzing problems.

This patch adds saving log files before deleting containers. The log files are
archived and saved to the /logs directory inside the container, which is a
volume mounted in the logs_cdw and logs_sdw1 directories for the corresponding
containers on the host.