Update approach to counting the number of rows for replicated tables at DML (#969)
Replicated tables are the same at all segments. Modifying replicated tables lead
to do the same processing at each segment. All segments calculate the count of
modified (updated, deleted or inserted) rows. The count on all segments will be
the same. It led to situation: QD got the same counts of modified rows from each
segments, summarized it, then divided it on number of segments. But it had to do
not in all cases. One such case: select from a function which contains DML on
replicated table. Such query led to an error: "consistency check on SPI tuple
count failed".

Also, the ExecutePlan function allows to limit the number of processing tuples,
but for DML operations the number of tuples is acquired from the sum of such
numbers from segments. It led to the same error, because QD processed the number
of tuples, which is less than all segments sent.

This patch modifies approach to counting the number of rows for replicated
tables at DML operations. Only one segment will calculate the count of modified
rows if the table is replicated. Other segments will send zero. It was done by
disabling the canSetTag flag for all segments except one. Dividing on number of
segments is no longer needed.

Also, for SPI calculating the number of processed tuples was added for cases,
when it is used at the _SPI_checktuples function. The behavior of _SPI_pquery
was changed to the behavior at PostgreSQL.

NOTE: the canSetTag flag enables saving the lastTid variable for the INSERT
queries. But it is useless, because it is used at only one place: at the
currtid_byreloid function, which has been never used. It was removed by
PostgreSQL 14:
https://github.com/postgres/postgres/commit/7b94e999606e2e2e10d68d544d49fc5a5d5785ac

(cherry picked from commit 08a4193d46b7fe8667feaf61b0b15871bcb97fc0)