Add dockerfile which is based on Ubuntu 22.04
There is no sudo in the ubuntu:22.04 image, so sudo is moved from
README.ubuntu.bash to README.linux.md. Add packages, which are required to build
GPDB, to README.ubuntu.bash. The new Dockerfile.ubuntu file is similar to
arenadata/Dockerfile. The first RUN setups the environment to run
compile_gpdb.bash, the last - to run regression tests.