Add dockerfile which is based on Ubuntu 22.04 (#1009)
There is no sudo in the ubuntu:22.04 image, so sudo is moved from
README.ubuntu.bash to README.linux.md. Add packages, which are required to build
and run GPDB, to README.ubuntu.bash. The new Dockerfile.ubuntu file is similar
to arenadata/Dockerfile. The first RUN setups the environment to run
compile_gpdb.bash and contains all changes to the ubuntu:22.04 image, which can
be made without the source code. So this layer will not be rebuilt for every
commit.