Disable GUC gp_log_stack_trace_lines by default
Gpdb uses addr2line to add file names and line numbers to the stack trace. But
addr2line can use a large amount of memory, the consumption of which we cannot
control, which can lead to OOM on the entire host.
This behavior is controlled by GUC gp_log_stack_trace_lines which is true by
default.

This patch changes the default value for gp_log_stack_trace_lines to false so as
not to run addr2line every time the stack is logged.