Decouple Orca from Explain component (#1124)
Problem:
Explain component had compile-time dependencies on Orca. It introduced
difficulties for moving Orca's functionality into a shared lib.

This patch:
1. Adds explain hook implementation to the gp_orca extension. It handles DXL
output format of explain command.
2. Removes all Orca related code from explain component in the GPDB core.
3. Adds function standard_ExplainOneQuery() that now contains the default
explain code (thus it can be invoked from the hook).
4. Removes 'planGen' field from 'PlannedStmt' structure. Its type was enum, and
it had only 2 values: Planner and Optimizer (for Postgres planner and Orca
planner respectively). It was not suitable for a general planner plugged in from
an extension.
5. Adds 'plannerName' field to 'PlannedStmt' structure. It is now used to output
the name of the external planner, if it is used.
6. Updates 'compute_jit_flags()' due to changes above. Now gp_orca extension
defines its own implementation for computing JIT parameters. Orca related logic
for JIT parameters is removed from the core.
7. Adds new ABI ignore file, as the 'PlannedStmt' structure is updated.