Rework gp_optimizer_functions.c to use runtime check of Orca's presence (#1126)
Problem:
Functions in gp_optimizer_functions.c had a compile-time dependency on Orca.
It introduced difficulties for moving Orca's functionality into a shared lib.

This patch:
1. Moves implementation of 'DisableXform()', 'EnableXform()',
'LibraryVersion()' into gp_orca shared lib.
2. Updates implementation of wrapper functions 'enable_xform()', 
'disable_xform()', 'gp_opt_version()'. Now they try to load the underlying 
functions from a shared lib. Attempt to load a symbol from a shared lib is 
wrapped into PG_TRY & PG_CATCH (with all necessary shenanigans), as the lib may 
be missing. In this case we catch the error, and show an appropriate message.
3. Updates tests due to change of the message if Orca is not enabled.

Note: we do not move the wrapper's implementation into gp_orca extension, as
they are already a part of the system catalog. And we'd like to avoid changing
the system catalog.