Move setting of 'optimizer' GUC to the extension (#1128)
Problem:
File 'guc_gp.c' contained a compile-time dependency on Orca, related to
'optimizer' GUC. It prevented moving of all Orca's code into a shared lib and
making its work transparent to the GPDB core.

This patch:
1. Sets the default value for 'optimizer' GUC to 'false'.
2. Adds enabling of 'optimizer' GUC in gp_orca extension shared lib.
3. Updates check in 'check_optimizer()' function. Now it utilizes runtime check
of planner_hook presence to distinguish if an external planner is available or
not.

Note: the semantic of 'optimizer' GUC has changed. Previously it was used
solely to enable & disable Orca. Now it is used to enable & disable any general
external planner. The name of GUC hasn't been changed, because too many places
in code and infrastructure rely on it, and we aim not to blow up the size of
the patch.