ADBDEV-6625 Decouple ORCA from the pg_hint_plan extension (#1127)
Decouple ORCA from the pg_hint_plan extension.

This change moves the definition of plan_hint_hook from ORCA, which is going to
be an extension to the planner. Although this hook is currently used only in
ORCA, it makes it possible to use it with any planner that needs to get a hint
list from an extension, such as pg_hint_plan.

Additionally, we make this hook, plan_hint_hook, typed and return a pointer to
HintState, as it actually does. There is no reason to keep it generic and
returning void *.