Fix EnableXform() & DisableXform() functions when optimizer is off (#1135)
Problem:
If a session was started with disabled 'optimizer' GUC, and EnableXform() or
DisableXform() was called from a query, the query failed with a crash. The issue
was found on 'rowhints' test.

Cause:
The functions mentioned above tried to access internal Orca's structures, that
were not initialized, as the init was done during first planning only if
'optimizer' was on.

Fix:
Init Orca when planner is invoked first time regardless the value of
'optimizer'.

Note:
One can think that it is better to perform Orca init somewhen at the backend
init. But we'd like to keep Orca init after GPMemoryProtect_Init() (as it was
before Orca moved to an extension). And there is no existing standard hook to
call after GPMemoryProtect_Init() and before PostgresMain() enters its main
loop. Thus it is where it is.

No new tests are added, as existing 'rowhints' test can catch this issue.