Fix gpfdist work with openssl 3.0
Openssl 3.0 does not allow to use self-signed certificates by default. It should
be created with the "-verify" flag. This flag verifies the self-signature on the
request. Also self-signed certificates does not work with host "127.0.0.1".

Patch add the "-verify" flag to command, which generates certificates, and
changes host "127.0.0.1" to "localhost" in gpfdist's ssl test. Such changes
successfully work with openssl 1.1.1.