Move Orca src files into the extension (#1151)
As a part of the bigger task of separating Orca into a Postgres extension, this
patch:
1. Moves 'src/backend/gporca/' and 'src/backend/gpopt/' folders with source
files into the extension.
2. Moves related headers into the extension.
3. Moves 'src/backend/optimizer/plan/orca.c' into the extension. As the
extension already contains 'orca.c' file with hooks implementation, old 'orca.c'
is renamed to 'orca_entry.c'.
4. All obj file names of newly added files are added into OBJS of
'gpcontrib/orca/Makefile'. It is mandated by the build infrastructure for
extensions (called PGXS). Old makefiles in 'gporca' and 'gpopt' are removed, as
no more needed. But CMake files are preserved, allowing to launch 'gporca' unit
tests.
5. Updates ABI ignore files, as some symbols are now missing in the core.
6. Moves OptimizerMemoryContext into the extension, as now it is referenced only
by the code in the extension.
7. Updates 'fmt' and 'tidy' tools and moves them into the extension, as they
were used only for Orca code.
Note: 'src/tools/vagrant' also contains some references to Orca code, but it
looks too obsolete to be workable (as it references the old gporca git repo,
when Orca wasn't a part of the core). So it is left unchanged.
8. Updates Readme and infrastructure files due to all changes above.