Fix 'gpconfig' behave test (#1158)
Problem:
After recent changes related to moving Orca into a shared lib, 'gpconfig'
behave test started to fail.

Cause:
Fail could be reproduced with the following commands:
"gpconfig -c optimizer -v on"
"gpstop -ar"

'gpconfig' alters 'postgresql.conf' by adding 'optimizer = on' there. So, the
issue appeared when we tried to set 'optimizer' GUC from the 'postgresql.conf'
file. At this moment shared preload libraries are yet not processed, so the
planner hook is yet not registered. It led to an error in 'check_optimizer()'
validation function when setting 'optimizer'.

Fix:
In case planner hook is not registered, issue an error in 'check_optimizer()'
only if source is more than PGC_S_ARGV. So, for cases when 'optimizer' is set
from the config file, env variable or postmaster's cmd argument, we allow
setting 'optimizer' to 'on' even if planner hook is yet not registered, as it
can't be registered at this stage, only later.